[tm_unicode] Added tmu_printf, tmu_vprintf, tmu_fprintf, tmu_vfprintf. [tm_unicode] Added tests for console output. [tm_unicode] Added tmu_module_filename, tmu_module_directory, tmu_open_directory, tmu_close_directory, tmu_read_directory.

+++ src/tm_unicode/runtime_functions.h
@@ -112,6 +112,6 @@
 		#define TMU_STRCHRW tmu_strchrw
 	#endif
 
-	#define TMU_WCSLEN(str) lstrlenW((str))
+	#define TMU_WCSLEN(str) (size_t)lstrlenW((str))
 #endif /* defined(TMU_USE_WINDOWS_H) */
 /* clang-format on */
